+++ s4e-web/src/app/views/map-view/state/sentinel-search/sentinel-search.store.ts
@@ -11,6 +11,7 @@ export class SentinelSearchStore extends EntityStore<SentinelSearchState, Sentin
       loaded: false,
       metadataLoading: false,
       metadataLoaded: false,
+      showSearchResults: false,
       metadata: {
         common: {
           params: []