+++ s4e-web/src/environments/environment.hmr-no-persist.ts
@@ -11,8 +11,8 @@ export const environment = {
   ...sharedEnvironment,
   production: false,
   hmr: true,
-  hmrPersist: true,
-  inviteOnly: true,
+  hmrPersist: false,
+  inviteOnly: false,
   uiDevelopment: true
 };